% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multBinExact_2-3.R
\name{print.multfisher}
\alias{print.multfisher}
\title{Print Values from a \code{multfisher} Object}
\usage{
\method{print}{multfisher}(x, ...)
}
\arguments{
\item{x}{an object of class \code{multfisher}}

\item{...}{further arguments passed to other methods. Not used.}
}
\description{
Print the test results.
}
\seealso{
\code{\link{mfisher.test}}, \code{\link{plot.multfisher}}
}
\author{
Robin Ristl, \email{robin.ristl@meduniwien.ac.at}
}
